Move range code into a separate submodule
