+++ crates/core/src/world/generate/runoff/basin.rs
@@ -1,6 +1,6 @@
 use crate::{
     unwrap,
-    util::{Meter, Meter3},
+    util::unit::{Meter, Meter3},
     world::{
         generate::{
             runoff::pattern::{RunoffDestination, RunoffPattern},
